fetch/metadata/generated/element-picture.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-picture.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, sourceEl, sourceAttr, attributes) {
    const picture = document.createElement('picture');
    const els = {
      img: document.createElement('img'),
      source: document.createElement('source')
    };
    picture.appendChild(els.source);
    picture.appendChild(els.img);
    document.body.appendChild(picture);

    for (const [ name, value ] of Object.entries(attributes)) {
      els.img.setAttribute(name, value);
    }

    return new Promise((resolve) => {
        els[sourceEl].setAttribute(sourceAttr, url);
        els.img.onload = els.img.onerror = resolve;
      });
  }

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - img[src] - Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite']),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[src] - Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite']),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - img[srcset] - Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite']),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - source[srcset] - Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - img[src] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - img[srcset] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - source[srcset] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {"crossorigin": ""}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[src] - attributes: crossorigin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {"crossorigin": ""}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[srcset] - attributes: crossorigin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {"crossorigin": ""}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - source[srcset] - attributes: crossorigin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {"crossorigin": "anonymous"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[src] - attributes: crossorigin=anonymous');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {"crossorigin": "anonymous"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[srcset] - attributes: crossorigin=anonymous');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {"crossorigin": "anonymous"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - source[srcset] - attributes: crossorigin=anonymous');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {"crossorigin": "use-credentials"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[src] - attributes: crossorigin=use-credentials');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {"crossorigin": "use-credentials"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - img[srcset] - attributes: crossorigin=use-credentials');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {"crossorigin": "use-credentials"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - source[srcset] - attributes: crossorigin=use-credentials');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - img[src] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - img[srcset] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - source[srcset] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - img[src] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'img',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - img[srcset] - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'source',
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - source[srcset] - no attributes');
  </script>
  </body>
</html>
